Split APIv3 tests on different files
